+++ rustfst/src/lib.rs
@@ -142,3 +142,10 @@ mod parsers;
 
 /// A representable float near .001. (Used in Quantize)
 pub(crate) const KDELTA: f32 = 1.0f32 / 1024.0f32;
+
+pub mod prelude {
+    pub use crate::fst_traits::*;
+    pub use crate::fst_impls::*;
+    pub use crate::semirings::*;
+    pub use crate::algorithms::*;
+}